// src/types.ts
export interface Movie {
    Title: string;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
    Rated?: string;
    Released?: string;
    Runtime?: string;
    Genre?: string;
    Director?: string;
    Writer?: string;
    Actors?: string;
    Plot?: string;
    Language?: string;
    Country?: string;
    Awards?: string;
    Metascore?: string;
    imdbRating?: string;
    imdbVotes?: string;
    Production?: string;
    Website?: string;
    Response?: string;
  }
  